+++ packages/jsts/src/analysis/projectAnalysis/analyzeWithWatchProgram.ts
@@ -49,16 +49,14 @@ export async function analyzeWithWatchProgram(
     const filenames = options.rootNames;
     for (const filename of filenames) {
       // only analyze files which are requested
-      if (files[filename]) {
-        results.files[filename] = analyzeFile(
-          {
-            filePath: filename,
-            fileContent: files[filename].fileContent ?? (await readFile(filename)),
-            fileType: files[filename].fileType,
-            tsConfigs: [tsConfig],
-          },
-          files[filename].language ?? DEFAULT_LANGUAGE,
-        );
+      if (files[filename] && pendingFiles.has(filename)) {
+        results.files[filename] = analyzeFile({
+          filePath: filename,
+          fileContent: files[filename].fileContent ?? (await readFile(filename)),
+          fileType: files[filename].fileType,
+          language: files[filename].language ?? DEFAULT_LANGUAGE,
+          tsConfigs: [tsConfig],
+        });
         pendingFiles.delete(filename);
       }
     }
